Adding JS mixin support for QtWebEngine and QtWebKit (so either one will be found and will work)
