+++ src/timeline/index.html
@@ -17,7 +17,7 @@
         <script type="text/javascript" src="media/js/angular-animate.min.js"></script>
 		
         <!-- OpenShot JavaScript Sources -->
-        <script type="text/javascript" src="js/qwebchannel.js"></script>
+        {{MIXIN_JS_INCLUDE}}
         <script type="text/javascript" src="app.js"></script>
         <script type="text/javascript" src="js/functions.js"></script>
     	<script type="text/javascript" src="js/controllers.js"></script>
